Put extracting and compilation of stemming table into one script.
